requirements.txt
torch
transformers
sentencepiece
torchaudio
librosa
opencv-python
imageio-ffmpeg
scikit-learn
numpy
pandas
tqdm
matplotlib
seaborn
pyyaml
